chore: Move emulator tests into a new file
